requirements.txt
absl-py==2.3.1
accelerate==0.21.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
backoff==2.2.1
bert-score==0.3.13
bitsandbytes==0.41.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
datasets==4.1.1
debugpy==1.8.16
decorator==5.2.1
deepspeed==0.14.5
dill==0.4.0
distro==1.9.0
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.117.1
ffmpy==0.6.1
filelock==3.19.1
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
gradio==4.16.0
gradio_client==0.8.1
h11==0.14.0
hf-xet==1.1.10
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.35.1
idna==3.10
importlib_resources==6.5.2
ipykernel==6.30.1
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1756321760188/work
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
latex2mathml==3.78.1
linkify-it-py==2.0.3
-e git+https://github.com/haotian-liu/LLaVA.git@c121f0432da27facab705978f83c4ada465e46fd#egg=llava
-e git+https://github.com/microsoft/LLaVA-Med.git@30697ca50b5c29a8e955c99330b259776aef27b9#egg=llava_med
lxml==6.0.2
markdown-it-py==2.2.0
markdown2==2.5.4
MarkupSafe==2.1.5
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
multiprocess==0.70.16
narwhals==2.5.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.13.0
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
openai==1.12.0
orjson==3.11.3
packaging==25.0
pandas==2.3.2
parso==0.8.5
peft==0.6.0
pexpect @ file:///croot/pexpect_1750072593878/work
pillow==10.4.0
platformdirs==4.4.0
portalocker==3.2.0
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.1
psutil==7.0.0
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-triton==3.1.0+cf34004b8a
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.0.2
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rich==14.1.0
rouge_score==0.1.2
rpds-py==0.27.1
ruff==0.13.2
sacrebleu==2.5.1
safetensors==0.6.2
scikit-learn==1.2.2
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy-split_1739790642651/work/dist/scipy-1.15.2-cp310-cp310-linux_x86_64.whl#sha256=9e52bad6c3294d1a5b04a13632118ca2157130603c6c018c2d710162b223b27e
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.39.0
shellingham==1.5.4
shortuuid==1.0.13
six @ file:///croot/six_1744271502820/work
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.48.0
svgwrite==1.4.3
sympy==1.13.1
tabulate==0.9.0
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1741878222898/work
tiktoken==0.11.0
timm==0.6.13
tokenizers==0.15.1
tomlkit==0.12.0
torch==2.1.2
torchvision==0.16.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.37.2
triton==2.1.0
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.5.0
uvicorn==0.37.0
wandb==0.22.1
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==11.0.3
xxhash==3.5.0
yarl==1.20.1
